requirements.txt
absl-py==0.1.10
bleach==1.5.0
cv2wrap==1.0
cycler==0.10.0
decorator==4.2.1
futures==3.1.1
h5py==2.7.1
html5lib==0.9999999
imgaug==0.2.5
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.1.3
Markdown==2.6.11
matplotlib==2.1.2
networkx==2.1
numpy==1.14.0
olefile==0.45.1
opencv-python==3.4.0.12
pandas==0.22.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
Pillow-SIMD==4.3.0.post0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.0
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
